Reverted back to old state.
To fully finish this game, I had to revert back to a codebase that I understood and wrote myself. I decided that to finish this game I want to learn to code it fully by myself by this point. I have used ChatGPT but understand all of the code that it wrote.